altered go parser to skip cs2 demos, will add this in after. Added new asset for spark job
