+++ dagster_workflow/demo-pipeline/demo_pipeline/assets/__init__.py
@@ -1,6 +1,7 @@
 from dagster import Definitions, load_assets_from_package_module
 
-from . import core
+from . import demo_collection, load_demo
 
-core_assets = load_assets_from_package_module(package_module=core, group_name="core")
+demo_collection_assets = load_assets_from_package_module(package_module=demo_collection, group_name="demo_collection")
+load_demo_assets = load_assets_from_package_module(package_module=load_demo, group_name="load_demo")
