+++ dagster_workflow/requirements.txt
@@ -5,5 +5,5 @@ pandas==1.1.5
 selenium==4.9
 webdriver-manager==3.8.6
 undetected-chromedriver==2.1.1
-dagit==1.3.13
-dagster==1.3.13
+dagit
+dagster==1.5.3
